client/requirements.txt
rich
prompt_toolkit
httpx
openai
google-generativeai
python-dotenv 